pyproject.toml
[project]
name = "threadsum"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "sentry-sdk>=2.56.0",
    "slack-bolt>=1.27.0",
]
